After much consulting and rigerous testing a Decision has been made.
Let it be known that on this day.and going forth we shall
continue to use the coc pugin in all it's third party ways

added the first keybinding just for shits and giggles
could be useful if I use it. (using ctrl-0 insert mod +one move)

the next is why it is going to stay for a while
TL/DR ..trial and lots of error got me these settings
got it to fuction well enough that I was using it
so after all that I'll keep it running.

mutipul tab work arounds found after checking into it